Add files via upload
Problem:
A phone company needs an application to keep track of the number of subscription(the number
of people and their names which chose the company) one has. Subscriptions can be added,
deleted and searched for. They can also be printed.
Explanation:
We are going to use a sorted map with the key the name of the person, which will be unique
and with the value the telephone number one has. The SortedMap will be implemented using a
binary search tree
ADT SORTED-MAP
DOMAIN:
 = {sm|sm is a sorted map with elements e =< k, v >, where k ∈ TKey and v ∈ TValue and rel ∈ Relation}

ADT SM ITERATOR
DOMAIN:
 = {it|it is an iterator over a sorted map sm with elements e =< k, v >, k ∈ TKey and v ∈ TValue and a rel∈ Relation }